Move away from troll package
