+++ eternalcore-core/src/main/java/com/eternalcode/core/feature/burn/BurnSettings.java
@@ -1,4 +1,4 @@
-package com.eternalcode.core.feature.troll.burn;
+package com.eternalcode.core.feature.burn;
 
 public interface BurnSettings {
     /**
